requirements.txt
selenium
haralyzer
pyyaml